fix: update imports after upstream/main rebase (PR #196 renamed engine-model → casehub-engine-common, moved CaseDefinitionRegistry/CaseInstanceCache)
Refs #195
