+++ casehub-blackboard/src/main/java/io/casehub/blackboard/subcase/SubCaseCompletionListener.java
@@ -20,7 +20,6 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.casehub.api.model.CaseStatus;
 import io.casehub.api.model.DefaultSubCaseCompletionStrategy;
 import io.casehub.api.model.SubCaseCompletionStrategy;
-import io.casehub.engine.internal.engine.cache.CaseInstanceCache;
 import io.casehub.engine.internal.event.CaseLifecycleEvent;
 import io.casehub.engine.internal.history.CaseHubEventType;
 import io.casehub.engine.internal.history.EventLog;
@@ -28,6 +27,7 @@ import io.casehub.engine.internal.history.EventStreamType;
 import io.casehub.engine.internal.model.CaseInstance;
 import io.casehub.engine.internal.work.CaseResumptionService;
 import io.casehub.engine.spi.EventLogRepository;
+import io.casehub.engine.spi.cache.CaseInstanceCache;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.event.ObservesAsync;
 import jakarta.inject.Inject;
