+++ casehub-blackboard/src/test/java/io/casehub/blackboard/subcase/SubCaseIntegrationTest.java
@@ -25,7 +25,7 @@ import io.casehub.api.model.CaseDefinition;
 import io.casehub.api.model.CaseStatus;
 import io.casehub.api.model.ContextChangeTrigger;
 import io.casehub.api.model.SubCase;
-import io.casehub.engine.internal.engine.cache.CaseInstanceCache;
+import io.casehub.engine.spi.cache.CaseInstanceCache;
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
